move most optimizer rules to new crate
