+++ datafusion/core/src/physical_plan/planner.rs
@@ -22,7 +22,7 @@ use super::{
     aggregates, empty::EmptyExec, expressions::binary, functions,
     hash_join::PartitionMode, udaf, union::UnionExec, values::ValuesExec, windows,
 };
-use crate::execution::context::{ExecutionProps, SessionState};
+use crate::execution::context::SessionState;
 use crate::logical_expr::utils::generate_sort_key;
 use crate::logical_plan::plan::{
     source_as_provider, Aggregate, EmptyRelation, Filter, Join, Projection, Sort,
@@ -53,7 +53,6 @@ use crate::physical_plan::windows::WindowAggExec;
 use crate::physical_plan::{join_utils, Partitioning};
 use crate::physical_plan::{AggregateExpr, ExecutionPlan, PhysicalExpr, WindowExpr};
 use crate::scalar::ScalarValue;
-use crate::variable::VarType;
 use crate::{
     error::{DataFusionError, Result},
     physical_plan::displayable,
@@ -63,6 +62,8 @@ use arrow::datatypes::{Schema, SchemaRef};
 use arrow::{compute::can_cast_types, datatypes::DataType};
 use async_trait::async_trait;
 use datafusion_expr::expr::GroupingSet;
+use datafusion_optimizer::ExecutionProps;
+use datafusion_optimizer::VarType;
 use datafusion_physical_expr::expressions::DateIntervalExpr;
 use datafusion_sql::utils::window_expr_common_partition_keys;
 use futures::future::BoxFuture;
