+++ datafusion/optimizer/src/common_subexpr_eliminate.rs
@@ -17,19 +17,21 @@
 
 //! Eliminate common sub-expression.
 
-use crate::error::Result;
-use crate::execution::context::ExecutionProps;
-use crate::logical_plan::plan::{Filter, Projection, Window};
-use crate::logical_plan::{
-    col,
-    plan::{Aggregate, Sort},
-    DFField, DFSchema, Expr, ExprRewritable, ExprRewriter, ExprSchemable, ExprVisitable,
-    ExpressionVisitor, LogicalPlan, Recursion, RewriteRecursion,
-};
-use crate::optimizer::optimizer::OptimizerRule;
+use crate::optimizer::OptimizerRule;
+use crate::ExecutionProps;
 use arrow::datatypes::DataType;
+use datafusion_common::{DFField, DFSchema, Result};
 use datafusion_expr::expr::GroupingSet;
-use datafusion_expr::utils::from_plan;
+use datafusion_expr::expr_rewriter::RewriteRecursion;
+use datafusion_expr::expr_visitor::Recursion;
+use datafusion_expr::{
+    col,
+    expr_rewriter::{ExprRewritable, ExprRewriter},
+    expr_visitor::{ExprVisitable, ExpressionVisitor},
+    logical_plan::{Aggregate, Filter, LogicalPlan, Projection, Sort, Window},
+    utils::from_plan,
+    Expr, ExprSchemable,
+};
 use std::collections::{HashMap, HashSet};
 use std::sync::Arc;
 
@@ -693,10 +695,12 @@ fn replace_common_expr(
 #[cfg(test)]
 mod test {
     use super::*;
-    use crate::logical_plan::{
-        avg, binary_expr, col, lit, sum, LogicalPlanBuilder, Operator,
+    use crate::test::test_table_scan;
+    use arrow::datatypes::{Field, Schema};
+    use datafusion_expr::logical_plan::table_scan;
+    use datafusion_expr::{
+        avg, binary_expr, col, lit, logical_plan::LogicalPlanBuilder, sum, Operator,
     };
-    use crate::test::*;
     use std::iter;
 
     fn assert_optimized_plan_eq(plan: &LogicalPlan, expected: &str) {
