+++ datafusion/optimizer/src/lib.rs
@@ -19,4 +19,29 @@
 //! some simple rules to a logical plan, such as "Projection Push Down" and "Type Coercion".
 
 #![allow(clippy::module_inception)]
-pub mod simplify_expressions;
+pub mod common_subexpr_eliminate;
+pub mod eliminate_filter;
+pub mod eliminate_limit;
+mod execution_props;
+pub mod filter_push_down;
+pub mod limit_push_down;
+pub mod optimizer;
+pub mod projection_push_down;
+pub mod single_distinct_to_groupby;
+pub mod subquery_filter_to_join;
+#[cfg(test)]
+pub mod test;
+pub mod utils;
+mod var_provider;
+
+pub use common_subexpr_eliminate::CommonSubexprEliminate;
+pub use eliminate_filter::EliminateFilter;
+pub use eliminate_limit::EliminateLimit;
+pub use execution_props::ExecutionProps;
+pub use filter_push_down::FilterPushDown;
+pub use limit_push_down::LimitPushDown;
+pub use optimizer::OptimizerRule;
+pub use projection_push_down::ProjectionPushDown;
+pub use single_distinct_to_groupby::SingleDistinctToGroupBy;
+pub use subquery_filter_to_join::SubqueryFilterToJoin;
+pub use var_provider::{VarProvider, VarType};
