+++ datafusion/optimizer/src/single_distinct_to_groupby.rs
@@ -17,13 +17,15 @@
 
 //! single distinct to group by optimizer rule
 
-use crate::error::Result;
-use crate::execution::context::ExecutionProps;
-use crate::logical_plan::plan::{Aggregate, Projection};
-use crate::logical_plan::ExprSchemable;
-use crate::logical_plan::{col, DFSchema, Expr, LogicalPlan};
-use crate::optimizer::optimizer::OptimizerRule;
-use datafusion_expr::utils::{columnize_expr, from_plan};
+use crate::optimizer::OptimizerRule;
+use crate::ExecutionProps;
+use datafusion_common::{DFSchema, Result};
+use datafusion_expr::{
+    col,
+    logical_plan::{Aggregate, LogicalPlan, Projection},
+    utils::{columnize_expr, from_plan},
+    Expr, ExprSchemable,
+};
 use hashbrown::HashSet;
 use std::sync::Arc;
 
@@ -200,9 +202,11 @@ impl OptimizerRule for SingleDistinctToGroupBy {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::logical_plan::{col, count, count_distinct, lit, max, LogicalPlanBuilder};
-    use crate::physical_plan::aggregates;
-    use crate::test::*;
+    use crate::test::test_table_scan;
+    use datafusion_expr::{
+        col, count, count_distinct, lit, logical_plan::LogicalPlanBuilder, max,
+        AggregateFunction,
+    };
 
     fn assert_optimized_plan_eq(plan: &LogicalPlan, expected: &str) {
         let rule = SingleDistinctToGroupBy::new();
@@ -311,7 +315,7 @@ mod tests {
                 vec![
                     count_distinct(col("b")),
                     Expr::AggregateFunction {
-                        fun: aggregates::AggregateFunction::Max,
+                        fun: AggregateFunction::Max,
                         distinct: true,
                         args: vec![col("b")],
                     },
